requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
ahocorapy==1.6.1
    # via -r requirements.in
amqp==5.0.6
    # via kombu
asgiref==3.4.1
    # via uvicorn
billiard==3.6.4.0
    # via celery
celery==5.1.2
    # via -r requirements.in
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
colorama==0.4.4
    # via
    #   tqdm
    #   uvicorn
fastapi==0.68.1
    # via -r requirements.in
future==0.18.2
    # via ahocorapy
greenlet==1.1.2
    # via sqlalchemy
h11==0.12.0
    # via uvicorn
httptools==0.2.0
    # via uvicorn
kombu==5.1.0
    # via celery
numpy==1.21.2
    # via pandas
pandas==1.3.3
    # via -r requirements.in
prompt-toolkit==3.0.20
    # via click-repl
pydantic==1.8.2
    # via
    #   fastapi
    #   sqlmodel
pymysql==1.0.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
python-dotenv==0.19.0
    # via
    #   -r requirements.in
    #   uvicorn
pytz==2021.3
    # via
    #   celery
    #   pandas
pyyaml==5.4.1
    # via uvicorn
redis==3.5.3
    # via -r requirements.in
requests==2.26.0
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
sqlalchemy==1.4.25
    # via sqlmodel
sqlalchemy2-stubs==0.0.2a17
    # via sqlmodel
sqlmodel==0.0.4
    # via -r requirements.in
starlette==0.14.2
    # via fastapi
tqdm==4.62.3
    # via -r requirements.in
typing-extensions==3.10.0.2
    # via
    #   pydantic
    #   sqlalchemy2-stubs
uvicorn[standard]==0.15.0
    # via -r requirements.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
watchgod==0.7
    # via uvicorn
wcwidth==0.2.5
    # via prompt-toolkit
websockets==10.0
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools

requests~=2.26.0
